Extract an atlas SDF renderer from the text layer renderer to prepare for a colored atlas
