+++ renderer/src/pipelines.rs
@@ -2,7 +2,7 @@ use crate::{
     bind_group_entries,
     pods::{TextureColorVertex, TextureVertex},
     primitives::Pipeline,
-    text_layer, texture,
+    texture,
     tools::{create_pipeline, BindGroupLayoutBuilder},
 };
 
@@ -11,7 +11,6 @@ pub fn create(
     device: &wgpu::Device,
     view_projection_bind_group_layout: &wgpu::BindGroupLayout,
     texture_bind_group_layout: &texture::BindGroupLayout,
-    text_layer_bind_group_layout: &text_layer::BindGroupLayout,
     shape_bind_group_layout: &wgpu::BindGroupLayout,
     targets: &[Option<wgpu::ColorTargetState>],
 ) -> Vec<(Pipeline, wgpu::RenderPipeline)> {
@@ -22,19 +21,6 @@ pub fn create(
         bind_group_layouts: &[view_projection_bind_group_layout, texture_bind_group_layout],
         push_constant_ranges: &[],
     });
-
-    let text_layer_shader =
-        &device.create_shader_module(wgpu::include_wgsl!("text_layer/text_layer.wgsl"));
-
-    let text_layer_pipeline_layout =
-        device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-            label: Some("Text Layer Pipeline Layout"),
-            bind_group_layouts: &[
-                view_projection_bind_group_layout,
-                text_layer_bind_group_layout,
-            ],
-            push_constant_ranges: &[],
-        });
 
     let shape_shader = &device.create_shader_module(wgpu::include_wgsl!("shape/shape.wgsl"));
 
@@ -69,18 +55,6 @@ pub fn create(
                 "fs_sdf_glyph",
                 &texture_vertex_layout,
                 &glyph_pipeline_layout,
-                targets,
-            ),
-        ),
-        (
-            Pipeline::TextLayer,
-            create_pipeline(
-                "Text Layer Pipeline",
-                device,
-                text_layer_shader,
-                "fs_sdf_glyph",
-                &text_layer_vertex_layout,
-                &text_layer_pipeline_layout,
                 targets,
             ),
         ),
