+++ renderer/src/text_layer/atlas_sdf/mod.rs
@@ -1,0 +1,7 @@
+//! An atlas based SDF renderer.
+
+mod bind_group;
+mod renderer;
+
+pub use bind_group::*;
+pub use renderer::*;
